Azure/functions/requirements.txt
# Azure Functions for Python
azure-functions>=1.18.0

# Azure SDK
azure-cosmos>=4.5.0
azure-storage-queue>=12.9.0
azure-identity>=1.15.0

# HTTP and RSS
feedparser>=6.0.10
requests>=2.31.0
aiohttp>=3.9.0

# AI/ML
anthropic>=0.18.0
# sentence-transformers>=2.3.0  # Too large for Azure Functions, not needed for MVP

# Data Processing
python-dateutil>=2.8.2
pytz>=2023.3

# Utilities
pydantic>=2.5.0
python-dotenv>=1.0.0

